kafka_producer/requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.9
prometheus-client==0.21.0
